requirements.txt
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.1
cryptography==35.0.0
Flask==2.0.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
msal==1.15.0
pycodestyle==2.7.0
pycparser==2.20
pyee==8.2.2
PyJWT==2.2.0
requests==2.26.0
slack-sdk==3.11.2
slackeventsapi==3.0.0
tabulate==0.8.9
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
